Create item.py
As of right now, this makes a really ugly csv with the item id, the name of the item, what appears to be an html formatted description that I could parse, but haven't done yet, and a plaintext description that doesn't exist for some items. Any of those can be removed from the output pretty easily if you just let me know.